Optimizing build outputs
