+++ workspaces/templates-lib/packages/template-ssr-server/.npmignore
@@ -1,0 +1,10 @@
+/*
+!/dist/
+!/bin/
+!/package.json
+!/README.md
+/dist/tsconfig.tsbuildinfo
+/**/*.spec.js
+/**/*.spec.js.map
+/**/*.spec.d.ts
+/**/*.spec.d.ts.map
