+++ workspaces/templates-lib/packages/template-ssr-server/README.md
@@ -1,0 +1,4 @@
+# Goldstack Server-Side Rendering Template Tools
+
+
+For more information, see [Server-Side Rendering Template](https://goldstack.party/templates/ssr).
