+++ workspaces/templates-lib/packages/template-ssr-server/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@goldstack/template-ssr-cli",
+  "name": "@goldstack/template-ssr-server",
   "version": "0.1.0",
-  "description": "CLI tools for server-side rendering template",
+  "description": "Building blocks for implementing server-side rendered pages.",
   "keywords": [
     "goldstack",
     "utility",
@@ -22,11 +22,7 @@
   "license": "MIT",
   "author": "Max Rohde",
   "sideEffects": false,
-  "main": "src/templateSSRCli.ts",
-  "bin": {
-    "template": "./bin/template",
-    "template-s3": "./bin/template"
-  },
+  "main": "src/templateSSRServer.ts",
   "scripts": {
     "build": "yarn clean && yarn compile",
     "build:watch": "yarn clean && yarn compile-watch",
@@ -35,6 +31,7 @@
     "compile-watch": "tsc -p tsconfig.json --watch",
     "compile-watch:light": "nodemon --watch ./src/ -e '*' --exec 'yarn compile'",
     "coverage": "jest --collect-coverage --passWithNoTests --config=./jest.config.js --runInBand",
+    "generate-schema": "ts-node scripts/generateSchemas.ts && cp schemas/* ../../../templates/packages/server-side-rendering/schemas",
     "prepublishOnly": "yarn run build",
     "publish": "utils-git changed --exec \"yarn npm publish $@\"",
     "test-ci": "jest --passWithNoTests --config=./jest.config.js --runInBand",
@@ -42,27 +39,30 @@
     "version:apply:force": "yarn version $@ && yarn version apply"
   },
   "dependencies": {
-    "@goldstack/template-ssr": "0.1.0",
-    "@goldstack/template-ssr-server": "0.1.0",
+    "@goldstack/infra": "0.3.37",
+    "@goldstack/infra-aws": "0.3.46",
     "@goldstack/utils-aws-lambda": "0.1.31",
-    "@goldstack/utils-cli": "0.2.14",
-    "@goldstack/utils-docker": "0.3.39",
-    "@goldstack/utils-log": "0.2.15",
+    "@goldstack/utils-esbuild": "0.4.2",
     "@goldstack/utils-package": "0.3.40",
-    "@goldstack/utils-package-config": "0.3.41",
+    "@goldstack/utils-package-config-embedded": "0.4.4",
     "@goldstack/utils-sh": "0.4.32",
     "@goldstack/utils-template": "0.3.36",
     "@goldstack/utils-terraform": "0.3.70",
-    "@goldstack/utils-terraform-aws": "0.3.69",
-    "source-map-support": "^0.5.21",
-    "yargs": "^17.5.1"
+    "@yarnpkg/esbuild-plugin-pnp": "^3.0.0-rc.11",
+    "esbuild": "^0.14.48",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "source-map-support": "^0.5.21"
   },
   "devDependencies": {
     "@goldstack/utils-docs-cli": "0.3.11",
     "@goldstack/utils-git": "0.1.35",
     "@goldstack/utils-package-config-generate": "0.2.16",
+    "@types/aws-lambda": "^8.10.88",
     "@types/jest": "^27.5.1",
     "@types/node": "^17.0.33",
+    "@types/react": "^16.9.49",
+    "@types/react-dom": "^18",
     "@types/yargs": "^17.0.10",
     "jest": "^28.1.0",
     "rimraf": "^3.0.2",
@@ -71,6 +71,6 @@
     "typescript": "^4.6.4"
   },
   "publishConfig": {
-    "main": "dist/src/templateSSRCli.js"
+    "main": "dist/src/templateSSRServer.js"
   }
 }
