+++ workspaces/templates-lib/packages/template-ssr-server/tsconfig.generate.json
@@ -1,0 +1,12 @@
+{
+  "extends": "./tsconfig.json",
+  "include": [
+    "./src/types/**/*"
+  ],
+  "compilerOptions": {
+    "incremental": false,
+    "declarationMap": false,
+    "outDir": "dist",
+    "baseUrl": "."
+  }
+}
