+++ workspaces/templates-lib/packages/template-ssr/package.json
@@ -41,17 +41,13 @@
   "dependencies": {
     "@goldstack/infra": "0.3.37",
     "@goldstack/infra-aws": "0.3.46",
+    "@goldstack/template-ssr-server": "0.1.0",
     "@goldstack/utils-aws-lambda": "0.1.31",
     "@goldstack/utils-esbuild": "0.4.2",
     "@goldstack/utils-package": "0.3.40",
     "@goldstack/utils-package-config-embedded": "0.4.4",
-    "@goldstack/utils-sh": "0.4.32",
     "@goldstack/utils-template": "0.3.36",
     "@goldstack/utils-terraform": "0.3.70",
-    "@yarnpkg/esbuild-plugin-pnp": "^3.0.0-rc.11",
-    "esbuild": "^0.14.48",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
     "source-map-support": "^0.5.21"
   },
   "devDependencies": {
@@ -61,8 +57,6 @@
     "@types/aws-lambda": "^8.10.88",
     "@types/jest": "^27.5.1",
     "@types/node": "^17.0.33",
-    "@types/react": "^16.9.49",
-    "@types/react-dom": "^18",
     "@types/yargs": "^17.0.10",
     "jest": "^28.1.0",
     "rimraf": "^3.0.2",
